% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{is an integer or a numeric double value with no decimals}

\item{y}{is an integer or a numeric double value with no decimals}
}
\value{
Function returns a numeric double value without decimals which corresponds to the greatest common divisor
}
\description{
This function calculates the greatest common divisor for the two given integer numbers
(or numeric double value with no decimals).
}
\details{
By consensus we considerate 0 to be the greatest common divisor of \code{x=0}, \code{y=0}
}
\examples{
euclidean(120,360) # = 120
euclidean(79,3) # = 1
euclidean(-8,36) # = 4
euclidean(85,0) # = 85
euclidean(-5L,20) # = 5
}
