% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3package-package.R
\docType{package}
\name{<lab3package>}
\alias{<lab3package>}
\title{lab3package}
\description{
This package has been built for the 3rd lab assignment of the "Advanced R Programming".
It contains 2 functions: \code{euclidean()} and \code{dijkstra()}.
}
\details{
Written by: Arash Haratian & Daniel Díaz-Roncero González
}
